Add recommendation to marketing server key
The campaignclassic.marketingServer key description mentions "marketing server", but that usually designates a URL that reaches only the main server of a Campaign installation. Therefore, setting this value would not use the full benefit of the infrastructure and would potentially overload the main server. A load-balanced URL is preferable, and the mirror page URL suits well for that purpose as it is normally load-balanced to Campaign servers that have database access, which is needed for registration.